refactor(tabs): separate tabs organization domain from adapters
