+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mark",
-  "version": "0.1.28",
+  "version": "0.1.29",
   "description": "Automated tab management",
   "repository": "https://github.com/bradcush/mark.git",
   "author": "Bradley Cushing <bradleycushing@gmail.com>",
