+++ src/manifest.json
@@ -1,7 +1,7 @@
 {
     "name": "Mark tab manager",
     "description": "The missing tab manager for Chrome",
-    "version": "0.1.28",
+    "version": "0.1.29",
     "manifest_version": 3,
     "action": {
         "default_title": "Mark"
