+++ src/tabs/sort.ts
@@ -1,17 +1,10 @@
 import { MkBrowser } from 'src/api/MkBrowser';
 import { makeSortName } from 'src/helpers/sortName';
 import { makeGroupName } from 'src/helpers/groupName';
-import { browser } from 'src/api/browser';
 import { logError, logVerbose } from 'src/logs/console';
 import { getStore } from 'src/storage/Store';
 import { categorize as categorizeTabs } from './categorize';
-
-/**
- * Compare to be used with name sorting
- */
-function compareNames(a: string, b: string) {
-    return a.localeCompare(b);
-}
+import { sort as sortBar } from './bar';
 
 /**
  * Separate grouped tabs from orphans
@@ -47,6 +40,32 @@ async function cluster(tabs: MkBrowser.tabs.Tab[]) {
 }
 
 /**
+ * Sort tabs alphabetically with nuance
+ */
+async function alphabetize(unsortedTabs: MkBrowser.tabs.Tab[]) {
+    logVerbose('alphabetize', unsortedTabs);
+    const { enableSubdomainFiltering } = await getStore().getState();
+    return unsortedTabs.sort((a, b) => {
+        const urlOne = a.url;
+        const urlTwo = b.url;
+        if (!urlOne || !urlTwo) {
+            throw new Error('No url for sorted tab');
+        }
+        const groupType = enableSubdomainFiltering ? 'granular' : 'shared';
+        const tabOneName = makeSortName({ type: groupType, url: urlOne });
+        const tabTwoName = makeSortName({ type: groupType, url: urlTwo });
+        return compareNames(tabOneName, tabTwoName);
+    });
+}
+
+/**
+ * Compare to be used with name sorting
+ */
+function compareNames(a: string, b: string) {
+    return a.localeCompare(b);
+}
+
+/**
  * Remove tabs that are pinned from the list
  */
 export function filter(tabs: MkBrowser.tabs.Tab[]): MkBrowser.tabs.Tab[] {
@@ -54,6 +73,29 @@ export function filter(tabs: MkBrowser.tabs.Tab[]): MkBrowser.tabs.Tab[] {
     const isTabPinned = (tab: MkBrowser.tabs.Tab) => !!tab.pinned;
     const nonPinnedTabs = tabs.filter((tab) => !isTabPinned(tab));
     return nonPinnedTabs;
+}
+
+/**
+ * Reorder browser tabs in the current
+ * window according to tabs list
+ */
+export async function render(tabs: MkBrowser.tabs.Tab[]): Promise<void> {
+    logVerbose('render', tabs);
+    // Not using "chrome.windows.WINDOW_ID_CURRENT" as we rely on real
+    // "windowId" in our algorithm which the representative -2 breaks
+    const staticWindowId = tabs[0].windowId;
+    const { forceWindowConsolidation } = await getStore().getState();
+    const tabsToUpdate = tabs.map(({ id }) => {
+        // Specify the current window as the forced window
+        const calculatedWindowId = forceWindowConsolidation
+            ? staticWindowId
+            : undefined;
+        return {
+            identifier: id,
+            windowId: calculatedWindowId,
+        };
+    });
+    sortBar(tabsToUpdate);
 }
 
 /**
@@ -76,62 +118,6 @@ export async function sort(
             : alphabetizedTabs;
     } catch (error) {
         logError('sort', error);
-        throw error;
-    }
-}
-
-/**
- * Sort tabs alphabetically with nuance
- */
-async function alphabetize(unsortedTabs: MkBrowser.tabs.Tab[]) {
-    logVerbose('alphabetize', unsortedTabs);
-    const { enableSubdomainFiltering } = await getStore().getState();
-    return unsortedTabs.sort((a, b) => {
-        const urlOne = a.url;
-        const urlTwo = b.url;
-        if (!urlOne || !urlTwo) {
-            throw new Error('No url for sorted tab');
-        }
-        const groupType = enableSubdomainFiltering ? 'granular' : 'shared';
-        const tabOneName = makeSortName({ type: groupType, url: urlOne });
-        const tabTwoName = makeSortName({ type: groupType, url: urlTwo });
-        return compareNames(tabOneName, tabTwoName);
-    });
-}
-
-/**
- * Reorder browser tabs in the current
- * window according to tabs list
- */
-export async function render(tabs: MkBrowser.tabs.Tab[]): Promise<void> {
-    try {
-        logVerbose('render', tabs);
-        // Not using "chrome.windows.WINDOW_ID_CURRENT" as we rely on real
-        // "windowId" in our algorithm which the representative -2 breaks
-        const staticWindowId = tabs[0].windowId;
-        const { forceWindowConsolidation } = await getStore().getState();
-        // We only care about catching errors with await in this case
-        // eslint-disable-next-line @typescript-eslint/no-misused-promises
-        tabs.forEach(async (tab) => {
-            const { id } = tab;
-            if (!id) {
-                throw new Error('No id for sorted tab');
-            }
-            const baseMoveProperties = { index: -1 };
-            // Specify the current window as the forced window
-            const staticWindowMoveProperties = {
-                windowId: staticWindowId,
-            };
-            // Current default uses the window for the current tab
-            const moveProperties = forceWindowConsolidation
-                ? { ...baseMoveProperties, ...staticWindowMoveProperties }
-                : baseMoveProperties;
-            // We expect calls to move to still run in parallel
-            // but await simply to catch errors properly
-            await browser.tabs.move(id, moveProperties);
-        });
-    } catch (error) {
-        logError('render', error);
         throw error;
     }
 }
